+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 #PWLocation Changelog
 
-##0.5.0 (Friday, June 18th, 2013)
+##0.8.0 (Tuesday, September 2nd, 2014)
+ * Added mock location manager for location testing
+ * Fixed several issues with the SLLocationManager callback events not firing
+
+##0.5.0 (Friday, June 18th, 2014)
  * Initial release (BETA)